refactor(frontend): extract PipelineTab (11/N)
Moves the kanban Pipeline tab + its PIPELINE_STAGES constant (only used
here) to tabs/PipelineTab.jsx. Props: t, applications, appLoading,
fetchApplications, setApplications.

App.jsx 2757 → 2654 LOC. Build clean.
